requirements.txt
Flask==0.10.1
SQLAlchemy==0.7.9
twilio==3.6.6
python-twitter